% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_operators.R
\name{.T}
\alias{.T}
\alias{.S}
\alias{.N}
\title{Fuzzy Logic Operators}
\usage{
.T(x, y, type = "Zadeh")

.S(x, y, type = "Zadeh")

.N(x, type = "Zadeh")
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{type}{Character string specifying which fuzzy logic family to refer to. (Refer to sets documentation (\code{\link[sets]{fuzzy}}) for more details.) Defaults to "Zadeh" type fuzzy logic.}
}
\value{
Returns a numeric vector, whose values depend on the fuzzy operator called (see Details for more information)
}
\description{
Alternative (non-dynamic) wrapper functions for \code{\link[sets:fuzzy]{sets::fuzzy_logic()}}. Here, operators are explicitly specified-- unlike the original wrapper functions in the \code{sets} package, their behaviour can be controlled with the an input parameter (\code{type}).
}
\details{
\describe{
\item{.T : conjunction operator}{
This function calculates the "t-norm", i.e. the equivalent of the "AND" operator in fuzzy logic.
It outputs the result of \code{sets::fuzzy_logic()$T}, but replaces values with 0 in the cases where values in either \code{x} or \code{y} is \code{NA} and the other is 0 (i.e., ignore the \code{NA} value).}
\item{.S : disjunction operator}{
This function calculates the "t-conorm", i.e. the equivalent of the "OR" operator in fuzzy logic.
It outputs the result of \code{sets::fuzzy_logic()$S}, but replaces values with 1 in the cases where values in either \code{x} or \code{y} is \code{NA} and the other is 1 (i.e., ignore the \code{NA} value).}
\item{.N : negation operator}{
This function calculates the equivalent of the "NOT" operator in fuzzy logic.
Outputs \code{1-x} for most fuzzy logic families, consistent with \code{sets::fuzzy_logic()$N}.}}
}
\section{References}{

Meyer D and Hornik K (2009). “Generalized and Customizable Sets in R.” Journal of Statistical Software, 31(2), pp. 1–27. doi: 10.18637/jss.v031.i02.
}

