% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzysys-package.r
\docType{package}
\name{fuzzysys}
\alias{fuzzysys}
\alias{fuzzysys-package}
\title{fuzzysys: Simple Fuzzy Inference and Fuzzy Systems}
\description{
Useful tools to build a simple fuzzy rule-based expert system model.
    Highly customizable for a variety of needs, without the clutter. A re-imagining of the fuzzy set and inference functions from the package `sets`, with more customizability and control.
}
\details{
The original `sets` package by David Meyer et al. can be found here: https://cran.r-project.org/web/packages/sets/index.html

Meyer D and Hornik K (2009). “Generalized and Customizable Sets in R.” Journal of Statistical Software, 31(2), pp. 1–27. doi: 10.18637/jss.v031.i02.
}
\author{
\strong{Maintainer}: Patricia Angkiriwang \email{p.angkiriwang@oceans.ubc.ca}

}
